backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-Babel==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
python-socketio==5.9.0
eventlet==0.33.3
requests==2.31.0
Pillow==10.1.0
ReportLab==4.0.7
scikit-learn==1.3.2
openai==0.28.1
web3==6.11.3
google-cloud-translate==3.12.1
cryptography==41.0.7
bcrypt==4.1.1
marshmallow==3.20.1
gunicorn==21.2.0